+++ B4.Api/Controllers/LkPruebas/CiclosPruebasController.cs
@@ -1,5 +1,5 @@
 using B4.Models.Entities.LkEntities;
-using B4.Models.Interfaces.LkInterfaces;
+using B4.Models.ServiceInterfaces;
 using Microsoft.AspNetCore.Mvc;
 
 namespace B4.Api.Controllers.Lk