+++ B4.Models/ServiceInterfaces/IDataBridgesFyBwService.cs
@@ -1,0 +1,9 @@
+using B4.Models.Entities.DataEntities;
+
+namespace B4.Models.ServiceInterfaces;
+
+public interface IDataBridgesFyBwService
+{
+    Task<DataBridgesFyBw?> GetByIdAsync(int id);
+    Task<IEnumerable<DataBridgesFyBw>> GetAllAsync();
+}